Separate system module from non_db_client
